FEV-695 Write RDT interpreter as soon as we show you the result (#1419)
A recent UI-stack optimization was to stop writing our data into the cloud all the time, and instead to only do so on navigation.  This leads to a situation where we can show you your RDT result, but never upload the fact that we did because you didn't navigate beyond that (e.g. you decided you got what you wanted out of the app, you backgrounded the app, or you even uninstalled the app).  The information about whether we've shown you an RDT is so important that we want to make sure we upload it.

So in uploaderMiddleware, we intercept the dispatch event that reflects that we've just shown you the interpretation, and force a data upload then.  This'll probably be one of the few exceptions to our general model of only uploading when you nav.