FEV-771 Fix timer regression (#1436)
https://github.com/AudereNow/learn/pull/1269 regressed the timers (no code review -- boo!).  I could have sworn I fixed the same issue in Fever;  or maybe someone fixed my same issue.

Anyway, the point is that it's easier to get timers right, IMO, if we always track the time remaining by calculating a difference between current time and original start time.  setTimeout and such can drift over time if the pattern is setTimeout -> do callback -> setAnotherTimeout many times in a row.

We now use State only to trigger rerendering when needed.  We use setInterval to keep a timer going, and we set/clear that interval on focus/blur.