Remove my face from homepage (#1483)
YouTube changed its policy to always require the profile photo of the person uploading a video to the site.  This made it so that my face would show up on the upper left of our site, on the video preview.

I've now uploaded the same video, but under the flu-support-au@auderenow.org account, which meants that it'll be the account, not mine, that owns the video.  We now also force the related videos to be from the same channel, not just whatever stuff comes after the video normally.