FEV-795: fix clientVersion to be object
In client code we set this as an object from buildInfo.json and TypeScript doesn't stop us.  All the Sniffles database records have objects here rather than string.  This ends up being an issue in Sniffles record recovery because I need to synthesize this field, and I want to make the records look the same as the other Sniffles records.