Make surveys remote-configurable (#1587)
We might not land HREC approval in time to send surveys out to people.  This adds a remote config to let us turn off survey-sending.

To minimize random changes to the codebase, this will simply skip the scheduling of the notification.  We make no attempt to remove the screen related to notifications, nor to not ask for notification permission.  This should be fine, in that we expect this to impact only users in the beginning week or so of our launch.