Custom build for Quidel's demo
Quidel has requested for their demo that:
* We skip screens related to asking the user for their interpretation
* We extend the timeout of RDT capture
* We make the result screen more specific to flu-a and flu-b test results

This PR makes these changes.  Note that we do not plan to commit this to the codebase, but instead just to build a custom build off this.

I've removed remote config checks related to Ubicomp, and instead defaulted those to true.  I've changed the flow to remove the user interpretation and high-contrast screens.