Implements phone login for EbPhotoStore (#1672)
* Implements phone login for EbPhotoStore

Adds Firebase auth, and introduces a modal dialog component that manages mobile login.  Firebase sends an SMS code, which you need to enter in the dialog in order to proceed.

We've introduced a test phone number in Firebase (via the Console) that's "2068675309" (from Tommy Tutone's "Jenny").  You can either enter that directly, with a validation code of 333333, or you can enter any one-digit number to get the same effect.

* Use setupBackInfo as a way to handle uniform logout