Introduce patient mapping in details (#1705)
Refactor SimpleMap to be embedable elsewhere by making it a normal React component, and add MapPage to be the /map component that renders that page.  Then refactor patient details to include the map.

SimpleMap now takes an array of EncounterDocuments, so that you can render a bunch of people or just one person.