Allow + symbols in phone numbers (#1708)
Firebase requires a full phone number in some form of internation phone number format, which requires a +.  We now change the keyboard to allow that.

We should later decide whether we want to auto-add country codes, etc.  For now, I think we're likely to disable SMS login for demo purposes, so I'll leave the bigger issue alone for now.