FEV-1042: retry cough import
Many of the previous import issues were around timeouts.  That is no longer happening, but we still see issues where the initial import fails because of auth and the subsequent auth always succeeds.

This adds a retry to the first firebase call in the import.  The hope is either to work around a transient issue, or to isolate the issue further.