Adds WhatsApp integration for EVD Track (#80)
This introduces two Cloud Functions from us:  one to intercept all WhatsApp messages to EVD Track's bot, and one to process Dialogflow's AI-driven conversation with the user.

We add two node modules:  dialogflow and dialogflow-fulfillment.

Photo storage isn't implemented yet, but is easy enough in a follow-on diff.  We store CHW information in the "whatsAppCHWs" collection so that we don't pollute EVD Track's mainline collections;  we store patient info in "whatsAppPatients" for similar reasons.  In a future diff, we can copy the relevant CHW information and patient information into the official EVD Track collections when we're ready for full integration.