Add Flows prototype app (#192)
In working with potential vNext partners, this is a basic project that introduces a React app that supports Firebase login.  It dynamically reloads the web app when you make changes and save;  it's linked with a new Firebase project for hosting and for cloud functions.  It supports email and Google authentication.

After this initial commit, more serious functionality will come.