terraform: add IPRD RDT-reader server in ECS
To support this change, I had to make our nginx image support SUBPATH and SERVERS variables.

SUBPATH is the URL sub-path to map to the service container.  E.g. for the RDT reader this is "/IPRD/rdt-reader/".

SERVERS specifies how to listen.  It is a space-separated list of entries where each entry looks like "protocol:inport:outport" (without the quotes).  For example, the entry "https:443:3000" species to listen on "https" port "443" and redirect to a service on port "3000".