Setting up install steps for FluApi in CircleCI (#563)
* Setting up install steps for FluApi in CircleCI

Adds jobs for building FluApi images and deploying them to ECS that can be invoked via the `build-service` script on developer workstations using a CircleCI API token. This should create a more repeatable and auditable deployment. All artifacts should be created in the same environment and we should have logs of deployments once they've been kicked off.

A few notes and more pressing concerns:

* Instead of tagging and deploying containers with a latest tag the script will explicitly use the commit as the tag. I think this will be convenient for determining which commit is deployed in prod and staging & better differentiates what is published.
* Build FluApi and deploy staging should probably be a single target. It is possible they will have different commits if an explicit commit is not provided.
* Since we're using CircleCI 2.0 we can't use orbs (CircleCI plug-ins) which would be helpful if we use CircleCI to automated AWS interactions. Unfortunately version 2.1 of the configuration removes the ability to invoke jobs the way we are doing. The new alternative, pipeline parameters & conditional workflows, is a notable change and relatively complicated in comparison.
* There are no checks in place presently about whether a given commit tag was already pushed to ECR. I believe trying to rebuild the same commit would result in a slightly unintuitive error since the user likely has no idea what commits are in ECR.
* There is an access key that has been generated with access to perform the requisite ECR actions. The details will need to be added to the CircleCI configuration.

Adding in notifications to Slack and pruning ECR images to a reasonable number seem like useful quality of life improvements to add on after we decide if running the deployment in CircleCI is right for us.

* Changing build option for staging to include deployment

Changing the name of the script to deploy-service and allowing us to specify build/publish/deploy in a single step.  Also a few quality of life improvements in error checking for the development script.