Enable end-to-end encryption of test result objects (1 of ~3) (#560)
First of approximately 3 commits that enable end-to-end encryption of
selected parts of medical record objects. This first part outlines the
API: clients of the encrypt module are expected to add an
EncryptionPolicy.MustEncrypt field to objects within a record they
wish to remain encrypted everywhere between client and final database.
So, for instance:

 testResults = {
   "date": ...,
   "hivStatus": {
      "encryptionPolicy": EncryptionPolicy.MustEncrypt,
      "positive": false,
      ...
    },
  }

Then later:

  applyEncryptionPolicy(testResults)

Only the objects marked for encryption are encrypted so as to
facilitate intermediate processing on non-encrypted nodes.

Part 2 of this diff series will plug in a real encryption
algorithm. (For now we just "encrypt" by converting strings to arrays
of UTF-8 bytes.)

Part 3 (and higher?) of the series will implement key management:
generation of keys within database and fetching thereof on clients.