FEV-1573: Migrating Slack notifier to a new project (#582)
* Migrating Slack notifier to a new project

The previous GitHub project we were using did not work well with the Node 10.x runtime in Lambda. Since we have a hard requirement to update our Lambda functions to Node 10.x or 12.x I've updated this to a new library. The newer library has a simpler deployment so I've modified init-local and removed the duplicate init-local from the cron Lambda module.

* Updating prod provider, cleaning up whitespace

* Correcting Slack notify project name