Attempt to save preview frames only from existing photos (#609)
If there is a survey document uploaded but the corresponding photo has not yet arrived, or will not arrive due to transport errors such as the user going offline etc, then we do not attempt to fetch that photo. It will be picked up when both the survey and photo exist.