integrate with the latest phase2 classifier
the latest phase2 classifier takes the test area image as the input and output the activation scores for the 5 output classes (both, flu-a, flu-b, invalid and negative). The prediction is the class with the highest score.

To make the demo available asap, this commit contains the minimal changes necessary to integrate the new classification model with the app. Further refactoring is necessary from a code perspective.